Fix Zabbix sessions table to clean
