+++ src/repositories/zabbix-host.ts
@@ -1,87 +1,10 @@
 import 'server-only'
-import axios, { AxiosRequestConfig } from 'axios'
 
 import { ZabbixGraph, ZabbixHost } from '@/common/interfaces/api/zabbix'
-import {
-  ZabbixBaseResponse,
-  ZabbixHostGraphResponse,
-} from '@/common/interfaces/api/zabbix-server'
+import { ZabbixHostGraphResponse } from '@/common/interfaces/api/zabbix-server'
+import { ZabbixRepository } from '@/repositories/zabbix'
 
-const { ZABBIX_URL, ZABBIX_USER, ZABBIX_PASS } = process.env
-
-export class ZabbixHostRepository {
-  /**
-   * POST request to Zabbix server
-   * @param method Zabbix API method
-   * @param params Zabbix API params
-   * @param auth Zabbix API auth token
-   * @returns Zabbix API result of response
-   * @throws Error if response is not valid
-   */
-  private static async post<T>(
-    method: string,
-    params: Record<string, unknown>,
-    auth?: string
-  ): Promise<T> {
-    const id = Math.floor(Math.random() * 114514) + 1
-    const url = `${ZABBIX_URL}/api_jsonrpc.php`
-    const headers: AxiosRequestConfig['headers'] = {
-      'Content-Type': 'application/json-rpc',
-    }
-    if (auth) {
-      headers.Authorization = `Bearer ${auth}`
-    }
-    const body = {
-      jsonrpc: '2.0',
-      method,
-      params,
-      id,
-    }
-
-    const res = await axios.post<ZabbixBaseResponse<T>>(url, body, {
-      headers,
-    })
-    if (res.data.error) {
-      throw new Error(res.data.error.message)
-    }
-    if (res.data.id !== id) {
-      throw new Error('ID mismatch')
-    }
-    if (!res.data.result) {
-      throw new Error('No result')
-    }
-
-    return res.data.result
-  }
-
-  /**
-   * Login to Zabbix server
-   * @returns Zabbix token
-   * @throws Error if login failed
-   */
-  private static async login(): Promise<string> {
-    const method = 'user.login'
-    const params = {
-      username: ZABBIX_USER,
-      password: ZABBIX_PASS,
-    }
-    const result = await this.post<string>(method, params)
-    return result
-  }
-
-  /**
-   * Logout from Zabbix server
-   * @param token Zabbix token
-   * @returns Is logout successful
-   * @throws Error if logout failed
-   */
-  private static async logout(token: string): Promise<boolean> {
-    const method = 'user.logout'
-    const params = {}
-    const result = await this.post<boolean>(method, params, token)
-    return result
-  }
-
+export class ZabbixHostRepository extends ZabbixRepository {
   /**
    * Zabbix hosts response to frontend hosts data
    * @param data Zabbix hosts response
@@ -132,14 +55,14 @@ export class ZabbixHostRepository {
   public static async getHosts(): Promise<ZabbixHost[]> {
     let token = ''
     try {
-      token = await this.login()
+      token = await this.loginApi()
 
       const method = 'host.get'
       const params = {
         output: ['hostid', 'name'],
         selectGraphs: ['graphid', 'name', 'graphtype'],
       }
-      const data = await this.post<ZabbixHostGraphResponse[]>(
+      const data = await this.postApi<ZabbixHostGraphResponse[]>(
         method,
         params,
         token
@@ -151,7 +74,7 @@ export class ZabbixHostRepository {
       throw new Error(`Failed to get hosts: ${e}`)
     } finally {
       if (token) {
-        await this.logout(token)
+        await this.logoutApi(token)
       }
     }
   }
